Fix I2C Pullup Issue
The combination of pullup resistors on the breakout, within the TXS0108, and on the BlackBoard Qwiic Level shifter conspired to make it impossible for the ICM to pull the data line low.

Fix is to rely on only the internal pullup resistors within the TXS0108.

Also removed a stray silk item from the board